[upgrade]赤外線操作機能追加
赤外線による家電の操作機能を追加と，それに伴う各種調整

+++ main.py
@@ -37,6 +37,8 @@ def main():
     try:
         data = '' # dataの初期化
         killword ='' # 前回認識した言葉を記憶するための変数
+        first_time = 'yes' #初回の認識をスルーするため
+        #time.sleep(3)
         while True:
 
             #print(data) # 認識した言葉を表示して確認
@@ -45,8 +47,13 @@ def main():
                 root = ET.fromstring('<?xml version="1.0"?>\n' + data[data.find('<RECOGOUT>'):].replace('\n.', ''))
                 for whypo in root.findall('./SHYPO/WHYPO'):
 
-                    word = whypo.get('WORD')# juliusで認識したWORDをwordに入れる
-                    if word == u'ロズビー':
+                    word = whypo.get('WORD') # juliusで認識したWORDをwordに入れる
+                    score = float(whypo.get('CM')) # 認識した単語の閾値
+
+                    print (word)
+                    print (score)
+
+                    if word == u'ロズビー' and score > 0.95:
                         #julius 停止
                         p.kill()
                         subprocess.call(["kill " + pid], shell=True)# juliusのプロセスを終了する。
@@ -55,17 +62,16 @@ def main():
                         #docomo_APIによる高精度認識
                         jtalk.responce('responce')
                         time.sleep(2)
-                        command = analysis.recognize()
+                        analysis.recognize()
 
                         #julius 再起動
                         p = subprocess.Popen([script], stdout=subprocess.PIPE, shell=True) # julius起動スクリプトを実行
                         pid = str(p.stdout.read().decode('utf-8')) # juliusのプロセスIDを取得
-                        time.sleep(1) # 3秒間スリープ
+                        time.sleep(1) # 1秒間スリープ
                         client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
                         client.connect((host, port)) #サーバーモードで起動したjuliusに接続
 
                     data = '' # dataの初期化
-
             else:
                 data += str(client.recv(1024).decode('utf-8')) #dataが空のときjuliusからdataに入れる
                 #print('NotFound')# juliusに認識する言葉がない。認識していない。
